Structured all source code
