+++ Pokemon-API/src/db/mock-pokemon.js
@@ -1,6 +1,5 @@
 const pokemons = [
     {
-     "id": 1,
      "name": "Bulbizarre",
      "hp": 25,
      "cp": 5,
@@ -9,7 +8,6 @@ const pokemons = [
      "created": new Date()
     },
     {
-     "id": 2,
      "name": "Salamèche",
      "hp": 28,
      "cp": 6,
@@ -18,7 +16,6 @@ const pokemons = [
      "created": new Date()
     },
     {
-     "id": 3,
      "name": "Carapuce",
      "hp": 21,
      "cp": 4,
@@ -27,7 +24,6 @@ const pokemons = [
      "created": new Date()
     },
     {
-     "id": 4,
      "name": "Aspicot",
      "hp": 16,
      "cp": 2,
@@ -36,7 +32,6 @@ const pokemons = [
      "created": new Date()
     },
     {
-     "id": 5,
      "name": "Roucool",
      "hp": 30,
      "cp": 7,
@@ -45,7 +40,6 @@ const pokemons = [
      "created": new Date()
     },
     {
-     "id": 6,
      "name": "Rattata",
      "hp": 18,
      "cp": 6,
@@ -54,7 +48,6 @@ const pokemons = [
      "created": new Date()
     },
     {
-     "id": 7,
      "name": "Piafabec",
      "hp": 14,
      "cp": 5,
@@ -63,7 +56,6 @@ const pokemons = [
      "created": new Date()
     },
     {
-     "id": 8,
      "name": "Abo",
      "hp": 16,
      "cp": 4,
@@ -72,7 +64,6 @@ const pokemons = [
      "created": new Date()
     },
     {
-     "id": 9,
      "name": "Pikachu",
      "hp": 21,
      "cp": 7,
@@ -81,7 +72,6 @@ const pokemons = [
      "created": new Date()
     },
     {
-     "id": 10,
      "name": "Sabelette",
      "hp": 19,
      "cp": 3,
@@ -90,7 +80,6 @@ const pokemons = [
      "created": new Date()
     },
     {
-     "id": 11,
      "name": "Mélofée",
      "hp": 25,
      "cp": 5,
@@ -99,7 +88,6 @@ const pokemons = [
      "created": new Date()
     },
     {
-     "id": 12,
      "name": "Groupix",
      "hp": 17,
      "cp": 8,
